1.servlet 修复BUG,多个tcp链接,使用同一个sessionID的数据无法共享

+++ src/main/java/com/github/netty/protocol/servlet/ServletHttpSession.java
@@ -64,11 +64,19 @@ public class ServletHttpSession implements HttpSession,Wrapper<Session>{
     }
 
     public void save(){
+	    if(id == null){
+		    return;
+	    }
         getServletContext().getSessionService().saveSession(unwrap());
+	    getServletContext().log("saveHttpSession : sessionId="+id);
     }
 
     public void remove(){
+    	if(id == null){
+    		return;
+	    }
         getServletContext().getSessionService().removeSession(getId());
+	    getServletContext().log("removeHttpSession : sessionId="+id);
     }
 
     @Override
